+++ taiga/projects/userstories/services.py
@@ -15,6 +15,7 @@ from contextlib import closing
 
 from django.conf import settings
 from django.db import connection
+from django.db.models import Q
 from django.utils import timezone
 from django.utils.translation import ugettext as _
 
@@ -25,6 +26,7 @@ from taiga.celery import app
 from taiga.events import events
 from taiga.projects.history.services import take_snapshot
 from taiga.projects.models import Project, UserStoryStatus, Swimlane
+from taiga.projects.milestones.models import Milestone
 from taiga.projects.notifications.utils import attach_watchers_to_queryset
 from taiga.projects.services import apply_order_updates
 from taiga.projects.tasks.models import Task
@@ -135,6 +137,129 @@ def reset_userstories_kanban_order_in_bulk(project: Project,
                               projectid=project.id)
 
 
+def update_userstories_backlog_or_sprint_order_in_bulk(user: User,
+                                                       project: Project,
+                                                       bulk_userstories: List[int],
+                                                       before_userstory: Optional[models.UserStory] = None,
+                                                       after_userstory: Optional[models.UserStory] = None,
+                                                       milestone: Optional[Milestone] = None):
+    """
+    Updates the order of the userstories specified adding the extra updates
+    needed to keep consistency.
+
+    Note: `after_userstory_id` and `before_userstory_id` are mutually exclusive;
+          you can use only one at a given request. They can be both None which
+          means "at the beginning of is cell"
+
+     - `bulk_userstories` should be a list of user stories IDs
+    """
+    # Get ids from milestones affected
+    milestones_ids = set(project.milestones.filter(user_stories__in=bulk_userstories).values_list('id', flat=True))
+    if milestone:
+        milestones_ids.add(milestone.id)
+
+    order_param = "backlog_order"
+
+    # filter user stories from milestone
+    user_stories = project.user_stories.all()
+    if milestone is not None:
+         user_stories = user_stories.filter(milestone=milestone)
+         order_param = "sprint_order"
+    else:
+         user_stories = user_stories.filter(milestone__isnull=True)
+
+    # exclude moved user stories
+    user_stories = user_stories.exclude(id__in=bulk_userstories)
+
+    # if before_userstory, get it and all elements before too:
+    if before_userstory:
+        user_stories = (user_stories.filter(**{f"{order_param}__gte": getattr(before_userstory, order_param)}))
+    # if after_userstory, exclude it and get only elements after it:
+    elif after_userstory:
+        user_stories = (user_stories.exclude(id=after_userstory.id)
+                                    .filter(**{f"{order_param}__gte": getattr(after_userstory, order_param)}))
+
+    # sort and get only ids
+    user_story_ids = (user_stories.order_by(order_param, "id")
+                                  .values_list('id', flat=True))
+
+    # append moved user stories
+    user_story_ids = bulk_userstories + list(user_story_ids)
+
+    # calculate the start order
+    if before_userstory:
+        # order start with the before_userstory order
+        start_order = getattr(before_userstory, order_param)
+    elif after_userstory:
+        # order start after the after_userstory order
+        start_order = getattr(after_userstory, order_param) + 1
+    else:
+        # move at the beggining of the column if there is no after and before
+        start_order = 1
+
+    # prepare rest of data
+    total_user_stories = len(user_story_ids)
+    user_story_orders = range(start_order, start_order + total_user_stories)
+
+    data = tuple(zip(user_story_ids,
+                     user_story_orders))
+
+    # execute query for update milestone and backlog or sprint order
+    sql = f"""
+    UPDATE userstories_userstory
+       SET {order_param} = tmp.new_order::BIGINT
+      FROM (VALUES %s) AS tmp (id, new_order)
+     WHERE tmp.id = userstories_userstory.id
+    """
+    with connection.cursor() as cursor:
+        execute_values(cursor, sql, data)
+
+    # execute query for update milestone for user stories and its tasks
+    bulk_userstories_objects = project.user_stories.filter(id__in=bulk_userstories)
+    bulk_userstories_objects.update(milestone=milestone)
+    project.tasks.filter(user_story__in=bulk_userstories).update(milestone=milestone)
+
+    # Generate snapshots for user stories and tasks and calculate if aafected milestones
+    # are cosed or open now.
+    if settings.CELERY_ENABLED:
+        _async_tasks_after_backlog_or_sprint_order_change.delay(bulk_userstories, milestones_ids, user.id)
+    else:
+        _async_tasks_after_backlog_or_sprint_order_change(bulk_userstories, milestones_ids, user.id)
+
+    # Sent events of updated stories
+    events.emit_event_for_ids(ids=user_story_ids,
+                              content_type="userstories.userstory",
+                              projectid=project.pk)
+
+    # Generate response with modified info
+    res = ({
+        "id": id,
+        "milestone": milestone.id if milestone else None,
+        order_param: order
+    } for (id, order) in data)
+    return res
+
+
+@app.task
+def _async_tasks_after_backlog_or_sprint_order_change(userstories_ids, milestones_ids, user_id):
+    try:
+        user = User.objects.get(id=user_id)
+    except User.DoesNotExist:
+        user = None
+
+    # Take snapshots for user stories and their taks
+    for userstory in  models.UserStory.objects.filter(id__in=userstories_ids):
+        take_snapshot(userstory, user=user)
+
+        for task in userstory.tasks.all():
+            take_snapshot(task, user=user)
+
+    # Check if milestones are open or closed after stories are moved
+    for milestone in  Milestone.objects.filter(id__in=milestones_ids):
+        _recalculate_is_closed_for_milestone(milestone)
+
+
+
 def update_userstories_kanban_order_in_bulk(user: User,
                                             project: Project,
                                             status: UserStoryStatus,
@@ -213,9 +338,9 @@ def update_userstories_kanban_order_in_bulk(user: User,
 
     # Update is_closed attr for user stories and related milestones
     if settings.CELERY_ENABLED:
-        _recalculate_is_closed_condition_and_take_snapshot.delay(bulk_userstories, user.id)
+        _async_tasks_after_kanban_order_change.delay(bulk_userstories, user.id)
     else:
-        _recalculate_is_closed_condition_and_take_snapshot(bulk_userstories, user.id)
+        _async_tasks_after_kanban_order_change(bulk_userstories, user.id)
 
     # Sent events of updated stories
     events.emit_event_for_ids(ids=user_story_ids,
@@ -233,7 +358,7 @@ def update_userstories_kanban_order_in_bulk(user: User,
 
 
 @app.task
-def _recalculate_is_closed_condition_and_take_snapshot(userstories_ids, user_id):
+def _async_tasks_after_kanban_order_change(userstories_ids, user_id):
     try:
         user = User.objects.get(id=user_id)
     except User.DoesNotExist:
@@ -344,15 +469,23 @@ def recalculate_is_closed_for_userstory_and_its_milestone(userstory):
         has_changed = open_userstory(userstory)
 
     if has_changed and userstory.milestone_id:
-        from taiga.projects.milestones import services as milestone_service
-
-        # Update is_close attr for the milestone
-        if milestone_service.calculate_milestone_is_closed(userstory.milestone):
-            milestone_service.close_milestone(userstory.milestone)
-        else:
-            milestone_service.open_milestone(userstory.milestone)
+        _recalculate_is_closed_for_milestone(userstory.milestone)
 
 
+def _recalculate_is_closed_for_milestone(milestone):
+    """
+    Check and update the open or closed condition for the milestone.
+
+    NOTE: [1] This method is useful if the userstory update operation has been done without
+              using the ORM (pre_save/post_save model signals).
+    """
+    from taiga.projects.milestones import services as milestone_service
+
+    # Update is_close attr for the milestone
+    if milestone_service.calculate_milestone_is_closed(milestone):
+        milestone_service.close_milestone(milestone)
+    else:
+        milestone_service.open_milestone(milestone)
 
 #####################################################
 # CSV