+++ taiga/projects/userstories/validators.py
@@ -92,51 +92,77 @@ class UserStoriesBulkValidator(ProjectExistsValidator, validators.Validator):
 
 # Order bulk validators
 
-class _UserStoryOrderBulkValidator(validators.Validator):
-    us_id = serializers.IntegerField()
-    order = serializers.IntegerField()
-
-
-class UpdateUserStoriesOrderBulkValidator(ProjectExistsValidator, validators.Validator):
+class UpdateUserStoriesBacklogOrderBulkValidator(ProjectExistsValidator, validators.Validator):
     project_id = serializers.IntegerField()
-    status_id = serializers.IntegerField(required=False)
     milestone_id = serializers.IntegerField(required=False)
-    bulk_stories = _UserStoryOrderBulkValidator(many=True)
-
-    def validate_status_id(self, attrs, source):
-        filters = {
-            "project__id": attrs["project_id"],
-            "id": attrs[source]
-        }
-
-        if not UserStoryStatus.objects.filter(**filters).exists():
-            raise ValidationError(_("Invalid user story status id. The status must belong "
-                                    "to the same project."))
-
-        return attrs
+    after_userstory_id = serializers.IntegerField(required=False)
+    before_userstory_id = serializers.IntegerField(required=False)
+    bulk_userstories = ListField(child=serializers.IntegerField(min_value=1))
 
     def validate_milestone_id(self, attrs, source):
-        filters = {
-            "project__id": attrs["project_id"],
-            "id": attrs[source]
-        }
+        milestone_id = attrs.get(source, None)
 
-        if not Milestone.objects.filter(**filters).exists():
-            raise ValidationError(_("Invalid milestone id. The milistone must belong to the "
-                                    "same project."))
+        if milestone_id:
+            filters = {
+                "project__id": attrs["project_id"],
+                "id": attrs[source]
+            }
+
+            if not Milestone.objects.filter(**filters).exists():
+                raise ValidationError(_("Invalid milestone id. The milestone must belong "
+                                        "to the same project."))
 
         return attrs
 
-    def validate_bulk_stories(self, attrs, source):
-        filters = {"project__id": attrs["project_id"]}
-        if "milestone_id" in attrs:
-            filters["milestone__id"] = attrs["milestone_id"]
+    def validate_after_userstory_id(self, attrs, source):
+        if attrs.get(source, None) is not None:
+            filters = {
+                "project__id": attrs["project_id"],
+                "id": attrs[source]
+            }
+            milestone_id = attrs.get("milestone_id", None)
+            if milestone_id:
+                filters["milestone__id"] = milestone_id
+            else:
+                filters["milestone__isnull"] = True
 
-        filters["id__in"] = [us["us_id"] for us in attrs[source]]
+            if not UserStory.objects.filter(**filters).exists():
+                raise ValidationError(_("Invalid user story id to move after. The user story must belong "
+                                        "to the same project and milestone."))
+
+        return attrs
+
+    def validate_before_userstory_id(self, attrs, source):
+        before_userstory_id = attrs.get(source, None)
+        after_userstory_id = attrs.get("after_userstory_id", None)
+
+        if after_userstory_id and before_userstory_id:
+            raise ValidationError(_("You can't use after and before at the same time."))
+        elif before_userstory_id is not None:
+            filters = {
+                "project__id": attrs["project_id"],
+                "id": attrs[source]
+            }
+            milestone_id = attrs.get("milestone_id", None)
+            if milestone_id:
+                filters["milestone__id"] = milestone_id
+            else:
+                filters["milestone__isnull"] = True
+
+            if not UserStory.objects.filter(**filters).exists():
+                raise ValidationError(_("Invalid user story id to move before. The user story must belong "
+                                        "to the same project and milestone."))
+
+        return attrs
+
+    def validate_bulk_userstories(self, attrs, source):
+        filters = {
+            "project__id": attrs["project_id"],
+            "id__in": attrs[source]
+        }
 
         if models.UserStory.objects.filter(**filters).count() != len(filters["id__in"]):
-            raise ValidationError(_("Invalid user story ids. All stories must belong to the same project "
-                                    "and, if it exists, to the same status and milestone."))
+            raise ValidationError(_("Invalid user story ids. All stories must belong to the same project."))
 
         return attrs